requirements.txt
streamlit
langchain==0.1.20
langchain-community==0.0.38
chromadb
pypdf
dashscope
langchain-dashscope
python-dotenv